refactor: Move Checksum class to app/models and enhance functionality
