+++ lib/storage_tables.rb
@@ -8,7 +8,6 @@ require "active_support/rails"
 require "storage_tables/version"
 require "storage_tables/engine"
 require "storage_tables/errors"
-require "storage_tables/checksum"
 
 require "marcel"
